@@ -1,9 +1,13 @@
 import { Router } from "express";
-const blogsController = require("../controllers/blogController");
+import { getBlogs, getBlog } from "../controllers/blogController";
+// import { authenticate } from "../middleware/auth/auth";
+// import { get } from "http";
 const router = Router();
 
-router.get("", blogsController.getBlogs);
+router.get("", getBlogs);
 
-router.get("/blog/:id", blogsController.getBlog);
+router.get("/blog/:id", getBlog);
+
+//router.get("", getBlogs);
 
 module.exports = router;
