@@ -1,11 +1,38 @@
 import { queryDB } from "../db/dataBase";
 import { Request, Response, NextFunction } from "express";
 
+interface RequestWithUser extends Request {
+  user: {
+    id: number;
+
+    // other properties...
+  };
+}
+
 async function getComments(req: Request, res: Response, next: NextFunction) {
-  const sql = "SELECT comments.*, users.username FROM comments JOIN users ON comments.user_id = users.user_id";
+  const sql =
+    "SELECT comments.*, users.username FROM comments JOIN users ON comments.user_id = users.id ORDER BY comments.blog_id ASC";
 
   try {
     const result = await queryDB(sql);
+    console.log(result, "getting all the comments");
+    res.status(200).json({ success: true, data: result });
+  } catch (err) {
+    res.status(500);
+    console.log(err);
+  } finally {
+    console.log("get all comments call over");
+  }
+}
+
+async function getCommentSpecificBlog(req: Request, res: Response, next: NextFunction) {
+  const blog_id = req.params.blog_id;
+  console.log(blog_id, "this is the blog id");
+  const sql =
+    "SELECT comments.*, users.username FROM comments JOIN users ON comments.user_id = users.id WHERE comments.blog_id = ?";
+
+  try {
+    const result = await queryDB(sql, [blog_id]);
     console.log(result, "getting all the comments");
     res.status(200).json({ success: true, data: result });
   } catch (err) {
@@ -33,12 +60,14 @@ async function getComment(req: Request, res: Response, next: NextFunction) {
   res.send("this is getting a single comment");
 }
 
-async function postComment(req: Request, res: Response, next: NextFunction) {
-  const { comment } = req.body;
-  const sql = "INSERT INTO comments (comment) VALUES (?)";
+async function postComment(req: any, res: Response, next: NextFunction) {
+  console.log(req.user_id, "user credentials");
+  const { comment, blog_id } = req.body;
+  const user_id = req.user_id;
+  const sql = "INSERT INTO comments (comment, user_id, blog_id, created_at) VALUES (?,?,?, NOW())";
 
   try {
-    const result = await queryDB(sql, [comment]);
+    const result = await queryDB(sql, [comment, user_id, blog_id]);
     console.log(result, "posting a new comment");
     res.status(200).json({ success: true, data: result });
   } catch (err) {
@@ -47,7 +76,6 @@ async function postComment(req: Request, res: Response, next: NextFunction) {
   } finally {
     console.log("post comment call over");
   }
-  res.send("this is posting a new comment");
 }
 
-export { getComments, getComment, postComment };
+export { getComments, getComment, postComment, getCommentSpecificBlog };
