@@ -5,4 +5,4 @@ function errorHandler(err: any, req: Request, res: Response, next: NextFunction)
   res.status(500).send("Something broke!");
 }
 
-module.exports = errorHandler;
+export { errorHandler };
