@@ -5,4 +5,4 @@ function logger(req: Request, res: Response, next: NextFunction) {
   next();
 }
 
-module.exports = logger;
+export { logger };
